backend/scripts/requirements.txt
# VTU Results Scraper - Required Packages
# Install with: pip install -r requirements.txt

# Database
mysql-connector-python>=8.0.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.0.0

# Image Processing & OCR
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Note: You also need to install Tesseract OCR separately:
# Download from: https://github.com/UB-Mannheim/tesseract/wiki
# For Windows: https://digi.bib.uni-mannheim.de/tesseract/tesseract-ocr-w64-setup-5.3.3.20231005.exe
